28-11

+++ src/Controller/ArticleController.php
@@ -11,6 +11,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\File\Exception\FileException;
 
 class ArticleController extends AbstractController
 {
@@ -25,6 +26,7 @@ class ArticleController extends AbstractController
     #[Route('/article/creer', name: 'app_article_create')]
     public function create(Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $article = new Article();
         $form = $this->createForm(ArticleType::class, $article);
         $form->handleRequest($request);
@@ -34,13 +36,28 @@ class ArticleController extends AbstractController
         //     ->setPublie(1)
         //     ->setDate(new DateTimeImmutable());
         //dd($article)
-        
-        if ($form->isSubmitted() && $form->isValid()) {
-            $article->setDate(new DateTimeImmutable());
-            $entityManager->persist($article);
 
+        if ($form->isSubmitted() && $form->isValid()) {
+            // Gestion de l'upload de l'image
+            $imageFile = $form->get('image')->getData();
+            if ($imageFile) {
+                $newFilename = uniqid() . '.' . $imageFile->guessExtension();
+                try {
+                    $imageFile->move(
+                        $this->getParameter('uploads_directory'),
+                        $newFilename
+                    );
+                    $article->setImage($newFilename);
+                } catch (FileException $e) {
+                    $this->addFlash('error', 'Erreur lors du téléchargement de l\'image.');
+                }
+            }
+
+            $article->setDate(new \DateTimeImmutable());
+            $entityManager->persist($article);
             $entityManager->flush();
-            $this->addFlash('success', 'Article Créer');
+
+            $this->addFlash('success', 'Article créé avec succès');
             return $this->redirectToRoute('app_article_list');
         }
 
@@ -55,7 +72,6 @@ class ArticleController extends AbstractController
     #[Route('/article/list', name: 'app_article_list')]
     public function list(EntityManagerInterface $entityManager): Response
     {
-        // Récupérer tous les articles de la base de données
         $articles = $entityManager->getRepository(Article::class)->findAll();
 
         return $this->render('article/list.html.twig', [
@@ -66,6 +82,7 @@ class ArticleController extends AbstractController
     #[Route('/article/update/{id}', name: 'app_article_edit')]
     public function update(Request $request, EntityManagerInterface $entityManager, int $id): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $article = $entityManager->getRepository(Article::class)->find($id);
 
         if (!$article) {
@@ -78,7 +95,7 @@ class ArticleController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->flush(); // Sauvegarde des modifications
+            $entityManager->flush();
     
             return $this->redirectToRoute('app_article_list');
         }
@@ -92,12 +109,20 @@ class ArticleController extends AbstractController
     #[Route('/article/delete/{id}', name: 'app_article_delete')]
     public function delete(EntityManagerInterface $entityManager, int $id): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $article = $entityManager->getRepository(Article::class)->find($id);
 
         if (!$article) {
             throw $this->createNotFoundException(
                 'No article found for id '.$id
             );
+        }
+
+        if ($article->getImage()) {
+            $imagePath = $this->getParameter('uploads_directory') . '/' . $article->getImage();
+            if (file_exists($imagePath)) {
+                unlink($imagePath);
+            }
         }
 
         $entityManager->remove($article);
